requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==44.0.3
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.7.0
filelock==3.18.0
Flask==3.1.0
flask-bcrypt==1.0.1
flask-cors==5.0.1
flask-session==0.8.0
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.3
greenlet==3.2.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pipenv==2025.0.2
platformdirs==4.3.8
playwright==1.52.0
pycparser==2.22
pydyf==0.11.0
pyee==13.0.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
pyphen==0.17.2
pytesseract==0.3.14
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
six==1.17.0
ssh-import-id==5.11
supervisor==4.2.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
virtualenv==20.31.2
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
fastapi==0.115.6
uvicorn==0.34.0

